+++ tests/test_hybracter.py
@@ -1,8 +1,9 @@
 import os
 import shutil
 import subprocess
-from pathlib import Path
 import unittest
+from pathlib import Path
+
 import pytest
 
 __author__ = "George Bouras"
@@ -21,7 +22,9 @@ db_dir: Path = "plassembler_db"
 def tmp_dir(tmpdir_factory):
     return tmpdir_factory.mktemp("tmp")
 
+
 threads = 2
+
 
 def remove_directory(dir_path):
     if os.path.exists(dir_path):
@@ -53,10 +56,34 @@ def test_hybracter_t_hybrid():
     remove_directory(outdir)
 
 
+def test_hybracter_t_hybrid_no_medaka():
+    """hybracter test-hybrid no medaka"""
+    outdir: Path = "test_hybracter_output"
+    cmd = f"hybracter test-hybrid --threads {threads} --output {outdir} --no_medaka"
+    exec_command(cmd)
+    remove_directory(outdir)
+
+
+def test_hybracter_t_hybrid_no_pypolca():
+    """hybracter test-hybrid no pypolca"""
+    outdir: Path = "test_hybracter_output"
+    cmd = f"hybracter test-hybrid --threads {threads} --output {outdir} --no_pypolca"
+    exec_command(cmd)
+    remove_directory(outdir)
+
+
 def test_hybracter_t_long():
     """hybracter test-long"""
     outdir: Path = "test_hybracter_output"
     cmd = f"hybracter test-long --threads {threads} --output {outdir}"
+    exec_command(cmd)
+    remove_directory(outdir)
+
+
+def test_hybracter_t_long_no_medaka():
+    """hybracter test-long no medaka"""
+    outdir: Path = "test_hybracter_output"
+    cmd = f"hybracter test-long --threads {threads} --output {outdir} --no_medaka"
     exec_command(cmd)
     remove_directory(outdir)
 
@@ -67,25 +94,52 @@ def test_hybracter_install():
     exec_command(cmd)
 
 
-@pytest.mark.dependency(depends=['test_hybracter_install'])
+"""
+test with csv
+"""
+
+
 def test_hybracter_hybrid_csv():
-    """test hybracter hybrid"""
+    """test hybracter hybrid default"""
     outdir: Path = "test_hybracter_output"
     input_csv: Path = test_data_path / "test_hybrid_input.csv"
     cmd = f"hybracter hybrid --input {input_csv} --threads {threads} --output {outdir} --databases {db_dir}"
     exec_command(cmd)
     remove_directory(outdir)
 
-@pytest.mark.dependency(depends=['test_hybracter_install'])
+
+def test_hybracter_no_pypolca_no_medaka():
+    """test hybracter hybrid no pypolca and no medaka"""
+    outdir: Path = "test_hybracter_output"
+    input_csv: Path = test_data_path / "test_hybrid_input.csv"
+    cmd = f"hybracter hybrid --input {input_csv} --threads {threads} --output {outdir} --no_pypolca --databases {db_dir} --no_medaka "
+    exec_command(cmd)
+    remove_directory(outdir)
+
+
+def test_hybracter_no_medaka():
+    """test hybracter hybrid no medaka"""
+    outdir: Path = "test_hybracter_output"
+    input_csv: Path = test_data_path / "test_hybrid_input.csv"
+    cmd = f"hybracter hybrid --input {input_csv} --threads {threads} --output {outdir} --databases {db_dir} --no_medaka"
+    exec_command(cmd)
+    remove_directory(outdir)
+
+
 def test_hybracter_no_pypolca():
     """test hybracter hybrid no pypolca"""
     outdir: Path = "test_hybracter_output"
     input_csv: Path = test_data_path / "test_hybrid_input.csv"
-    cmd = f"hybracter hybrid --input {input_csv} --threads {threads} --output {outdir} --no_pypolca --databases {db_dir}"
+    cmd = f"hybracter hybrid --input {input_csv} --threads {threads} --output {outdir} --no_pypolca --databases {db_dir} "
     exec_command(cmd)
     remove_directory(outdir)
 
-@pytest.mark.dependency(depends=['test_hybracter_install'])
+
+"""
+long
+"""
+
+
 def test_hybracter_long():
     """test hybracter long"""
     outdir: Path = "test_hybracter_output"
@@ -94,7 +148,21 @@ def test_hybracter_long():
     exec_command(cmd)
     remove_directory(outdir)
 
-@pytest.mark.dependency(depends=['test_hybracter_install'])
+
+def test_hybracter_long_no_medaka():
+    """test hybracter long"""
+    outdir: Path = "test_hybracter_output"
+    input_csv: Path = test_data_path / "test_long_input.csv"
+    cmd = f"hybracter long --input {input_csv} --threads {threads} --output {outdir} --databases {db_dir} --no_medaka"
+    exec_command(cmd)
+    remove_directory(outdir)
+
+
+"""
+single
+"""
+
+
 def test_hybracter_hybrid_single():
     """test hybracter hybrid single"""
     outdir: Path = "test_hybracter_output_hybr_single"
@@ -105,7 +173,19 @@ def test_hybracter_hybrid_single():
     exec_command(cmd)
     remove_directory(outdir)
 
-@pytest.mark.dependency(depends=['test_hybracter_install'])
+
+def test_hybracter_hybrid_single_no_medaka():
+    """test hybracter hybrid single no medaka"""
+    outdir: Path = "test_hybracter_output_hybr_single"
+    longreads: Path = test_data_path / "Fastqs/test_long_reads.fastq.gz"
+    reads1: Path = test_data_path / "Fastqs/test_short_reads_R1.fastq.gz"
+    reads2: Path = test_data_path / "Fastqs/test_short_reads_R2.fastq.gz"
+    cmd = f"hybracter hybrid-single -l {longreads} -1 {reads1} -2 {reads2} -c 50000 -s Sample1 --threads {threads} --output {outdir} --databases {db_dir} --no_medaka"
+    exec_command(cmd)
+    remove_directory(outdir)
+
+
+@pytest.mark.dependency(depends=["test_hybracter_install"])
 def test_hybracter_long_single():
     """test hybracter long"""
     outdir: Path = "test_hybracter_output_long_single"
@@ -114,7 +194,21 @@ def test_hybracter_long_single():
     exec_command(cmd)
     remove_directory(outdir)
 
-@pytest.mark.dependency(depends=['test_hybracter_install'])
+
+def test_hybracter_long_single_no_medaka():
+    """test hybracter long no medaka"""
+    outdir: Path = "test_hybracter_output_long_single"
+    longreads: Path = test_data_path / "Fastqs/test_long_reads.fastq.gz"
+    cmd = f"hybracter long-single -l {longreads} -c 50000 -s Sample1  --threads {threads} --output {outdir} --databases {db_dir} --no_medaka"
+    exec_command(cmd)
+    remove_directory(outdir)
+
+
+"""
+custom db
+"""
+
+
 def test_hybracter_dnaapler_custom_db():
     """test hybracter hybrid-single with custom_db"""
     outdir: Path = "test_hybracter_output_custom_db"
@@ -126,10 +220,12 @@ def test_hybracter_dnaapler_custom_db():
     exec_command(cmd)
     remove_directory(outdir)
 
+
 def test_citation():
     """test hybracter citation"""
     cmd = "hybracter citation"
     exec_command(cmd)
+
 
 def test_version():
     """test hybracter version"""
@@ -138,7 +234,6 @@ def test_version():
 
 
 class errors(unittest.TestCase):
-
     def test_no_db(self):
         """test hybracter with no db"""
         with self.assertRaises(RuntimeError):
@@ -148,5 +243,6 @@ class errors(unittest.TestCase):
             exec_command(cmd)
             remove_directory(outdir)
 
-# cleanup 
+
+# cleanup
 remove_directory(db_dir)